refactor(query): extract single-shot query into runner/single.rs
